% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_doc.R
\name{convert_doc}
\alias{convert_doc}
\title{Convert .doc-files containing evidence material to .txt-format}
\usage{
convert_doc(doc_folder)
}
\arguments{
\item{doc_folder}{Path to a certain folder containing .doc documents.}
}
\value{
A new folder was created containing the converted files. The name of this folder is the same as the folder containing the .doc files, with "_txt" appended to the end.
}
\description{
This function converts .doc files to .txt files using the LibreOffice program. Note: This function requires LibreOffice to be located in a folder named "Applications" and that you are currently using a Mac.
}
