% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_df.R
\name{create_df}
\alias{create_df}
\title{Create an article dataset from the .mlw-files that is ready for use with merge_df}
\usage{
create_df(mlw_folder)
}
\arguments{
\item{txt_folder}{Path to a certain folder containing .mlw-files.}
}
\value{
An evidence dataset containing the columns "lemma", "u*_bedeutung", "zitat" and "stelle" and that is ready for use with merge_df
}
\description{
This function creates a DataFrame from all .mlw-files located in the specified folder, containing the columns "lemma", "u*_bedeutung", "zitat" and "stelle". The files used in the example were automatically loaded with the package and should be located in your project folder as a folder named "toy", after you installed the package and executed convert_doc on the doc_folder. Please execute the example code manually and don't use the run_examples button.
}
\examples{
# Create the article dataFrame
toy_artikel <- create_df("toy/toy_mlw")
View(toy_artikel)
}
