% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_df_2.R
\name{create_df_2}
\alias{create_df_2}
\title{Create an Evidence-dataset from the .txt-files that is ready for use with merge_df}
\usage{
create_df_2(txt_folder)
}
\arguments{
\item{txt_folder}{Path to a certain folder containing .txt-files.}
}
\value{
An evidence dataset containing the columns "langer_beleg", "pruef_beleg" and "pruef_stelle" and that is ready for use with merge_df
}
\description{
This function creates a dataset from all .txt-files located in the specified folder, containing the columns "langer_beleg", "pruef_beleg" and "pruef_stelle". The folder used in the example below is available on the GitHub repository in a folder named "toy". You have to use convert_doc on the folder before executing the example code. The folder containing the .txt-files you want to use should be located in your project folder before using this function. Please execute the example code manually and don't use the run_examples button.For further informations see https://github.com/ATSzimmat/mlwdw26.
}
\examples{
# Create the Evidence-dataset
toy_belege <- create_df_2(txt_folder = "toy/toy_doc_txt")
# View the result
View(toy_belege)
}
