% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_doc.R
\name{convert_doc}
\alias{convert_doc}
\title{DOC-Dateien in TXT umwandeln via LibreOffice}
\usage{
convert_doc(doc_folder)
}
\arguments{
\item{doc_folder}{Pfad zum Ordner}
}
\description{
DOC-Dateien in TXT umwandeln via LibreOffice
}
