% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_doc.R
\name{convert_doc}
\alias{convert_doc}
\title{Convert the text from the evidence material to txt format}
\usage{
convert_doc(doc_folder)
}
\arguments{
\item{doc_folder}{Path to file with doc-documents}
}
\description{
Convert the text from the evidence material to txt format
}
