% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_doc.R
\name{convert_doc}
\alias{convert_doc}
\title{Convert .doc-files containing evidence material to .txt-format}
\usage{
convert_doc(doc_folder)
}
\arguments{
\item{doc_folder}{Path to a certain folder containing .doc-files.}
}
\value{
A new folder was created containing the converted files. The name of this folder is the same as the folder containing the .doc-files, with "_txt" appended to the end.
}
\description{
This function converts .doc-files to .txt files using the LibreOffice program.The folder used in the example below is available on the GitHub repository in a folder named "toy". The folder containing the .doc-files you want to convert should be located in your project folder before using this function. Note: This function requires LibreOffice to be located in a folder named "Applications" and that you are currently using a Mac. This is also the reason why the examples contain the comment "## Not run" - do not be confused by this - it merely serves to prevent R from causing errors. Unfortunately, this also means the "Run examples" button won't work - please run the example manually instead. For further informations see https://github.com/ATSzimmat/mlwdw26.
}
\examples{
\dontrun{# Convert the .doc-files from a example folder with .doc-files from the toy folder
convert_doc(doc_folder = "toy/toy_doc")
}
}
