% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_df.R
\name{create_df}
\alias{create_df}
\title{Create an Article-dataset from the .mlw-files that is ready for use with merge_df}
\usage{
create_df(mlw_folder)
}
\arguments{
\item{mlw_folder}{Path to a certain folder containing .mlw-files.}
}
\value{
An article dataset containing the columns "lemma", "u*_bedeutung", "zitat" and "stelle" and that is ready for use with merge_df
}
\description{
This function creates a dataset from all .mlw-files located in the specified folder, containing the columns "lemma", "u*_bedeutung", "zitat" and "stelle".The folder used in the example below is available on the GitHub repository in a folder named "toy". You should load the folder with the .mlw-files you want to use in your project folder before using this function. Please execute the example code manually and don't use the run_examples button. For further informations see https://github.com/ATSzimmat/mlwdw26.
}
\examples{
# Create the Article-dataset
toy_artikel <- create_df(mlw_folder = "toy/toy_mlw")
# View the result
View(toy_artikel)
}
