% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_df.R
\name{merge_df}
\alias{merge_df}
\title{Extend the quotes in the Article-dataset using the passages from the Evidence-dataset.}
\usage{
merge_df(lem_dat, bel_dat)
}
\arguments{
\item{lem_dat}{an article dataset created by create_df}

\item{bel_dat}{an evidence dataset created by create_df_2}
}
\value{
an article dataset with extended citations
}
\description{
This function takes an Article-dataset created by create_df and an Evidence-dataset created by create_df_2 and gives back an Article-dataset with extended quotes. The files used in the example below are available on the GitHub repository in a folder named "toy". Before executing the example code you should apply convert_doc, create_df and create_df_2 on the repective files. The Article-dataset generated by create_df and the Evidence-dataset you want to use, generated by convert_doc and create_df_2 should be loaded in your Global Environment before you use the function. For further informations see https://github.com/ATSzimmat/mlwdw26. All citations for which the function found no matching partner are saved in a CSV file, always named "fehler.csv". The source of the error that led to the missing partner is explicitly indicated. All quotes from the Article-dataset for which no suitable partner could be found are stored separately in a CSV-file named "fehler.csv", including metadata, and the conditions that were met or not met are explicitly stated. All quotes for which the function found multiple possible partners are saved in a CSV file, always named "nicht_eindeutig.csv".
}
\examples{
# Extend the Citations of the Article-dataset
final_df <- merge_df(lem_dat = toy_artikel , bel_dat = toy_belege)
# View the new Article-dataset
View(final_df)
# View the quotes with no suitable partner
errors <- readr::read_csv("fehler.csv")
View(errors)
# View the quotes with several possible partners
ambiguous <- readr::read_csv("nicht_eindeutig.csv")
View(ambiguous)

}
